Created share.buttons, import each button into this, output into share
